// src/utils/enums.ts
export enum MissionFileErrorType {
  NotFound = "notFound", // high-inferred; compactCamel; bundle hits=30
  ParseError = "parseError", // high-inferred; compactCamel; bundle hits=4
  PermissionError = "permission_error", // high-inferred; lower_snake; bundle hits=1
  ReadError = "readError", // high-inferred; compactCamel; bundle hits=15
}